feat: All Testability for Acting in one class Still not perfect and needs alot more finetuning. But selecting a testcase, a trajectory and which steeringcontroller to test seem to work already.
